LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-003-expected.txt


FAIL .target 1 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 2 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
PASS .target 3
FAIL .target 4 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 5 assert_equals:
<div class="target" data-expected-width="40"></div>
width expected 40 but got 0
FAIL .target 6 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 7 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 8 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
FAIL .target 9 assert_equals:
<div class="target" data-expected-width="30"></div>
width expected 30 but got 0
PASS .target 10
FAIL .target 11 assert_equals:
<div class="target" data-expected-width="20"></div>
width expected 20 but got 0
FAIL .target 12 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 13 assert_equals:
<div class="target" data-expected-width="50"></div>
width expected 50 but got 0
FAIL .target 14 assert_equals:
<div class="target" data-expected-width="50"></div>
width expected 50 but got 0
FAIL .target 15 assert_equals:
<div class="target" data-expected-width="70"></div>
width expected 70 but got 0
FAIL .target 16 assert_equals:
<div class="target" data-expected-width="70"></div>
width expected 70 but got 0
PASS .target 17
FAIL .target 18 assert_equals:
<div class="target" data-expected-width="20"></div>
width expected 20 but got 0
FAIL .target 19 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 20 assert_equals:
<div class="target" data-expected-width="110"></div>
width expected 110 but got 0
FAIL .target 21 assert_equals:
<div class="target" data-expected-width="110"></div>
width expected 110 but got 0
FAIL .target 22 assert_equals:
<div class="target" data-expected-width="110"></div>
width expected 110 but got 0
FAIL .target 23 assert_equals:
<div class="target" data-expected-width="100"></div>
width expected 100 but got 0
FAIL .target 24 assert_equals:
<div class="target" data-expected-width="20"></div>
width expected 20 but got 0
FAIL .target 25 assert_equals:
<div class="target" data-expected-width="20"></div>
width expected 20 but got 0
FAIL .target 26 assert_equals:
<div class="target" data-expected-width="120"></div>
width expected 120 but got 0
FAIL .target 27 assert_equals:
<div class="target" data-expected-width="110"></div>
width expected 110 but got 0
FAIL .target 28 assert_equals:
<div class="target" data-expected-width="100"></div>
width expected 100 but got 0
FAIL .target 29 assert_equals:
<div class="target" data-expected-width="100"></div>
width expected 100 but got 0
FAIL .target 30 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 31 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
PASS .target 32
FAIL .target 33 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 34 assert_equals:
<div class="target" data-expected-width="20"></div>
width expected 20 but got 0
FAIL .target 35 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 36 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 37 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 38 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
FAIL .target 39 assert_equals:
<div class="target" data-expected-width="10"></div>
width expected 10 but got 0
